+++ src/resources.hpp
@@ -78,6 +78,7 @@ SOFTWARE.
 #define PARAM_ERROR "error"
 #define PARAM_PLATFORM "platform"
 #define PARAM_BLE "ble"
+#define PARAM_ISPINDEL_CONFIG "ispindel-config"
 #define PARAM_HW_GYRO_READ_COUNT "gyro-read-count"
 // #define PARAM_HW_GYRO_READ_DELAY "gyro-read-delay"
 #define PARAM_HW_GYRO_MOVING_THREASHOLD "gyro-moving-threashold"