+++ src_docs/source/releases.rst
@@ -9,6 +9,7 @@ v1.4.0 (beta)
 * Changed wifi manager to modified version since the source project is no longer supported by author
 * Added possibility to set wifi settings in config (need manual reboot to apply)
 * Updated dependencies to latest versions as of Aug -23
+* Added feature to migrate formula and gyro calibration data from an existing ispindel configuration (don't do a full erase)
 
 v1.3.1
 ======